Restructure adapter-mongoose package
