+++ packages/adapter-mongoose/lib/tokenizers/index.js
