+++ packages/adapter-mongoose/lib/tokenizers/simple.js
